Fixed the following glitches: Glitch #1 1. open contour lines GUI. 2. click "Customize" in main GUI 3. customization GUI opens 4. don't do anything with it 5. now click on the main GUI again 6. the customization GUI disappears behind seadas 7. now click "Customize" 8. nothing happens, the customization GUI remains open but hidden behind seadas.
Glitch #2
1. click "Customize in main GUI
2. change a line color
3. click "Save"
4. change the Start Value
5. click "Customize" and the line color is back to default.

This makes sense, but the user may have edited names and colors of several levels.  These edits get lost and the user doesn't realize this.  If it were possible to determine whether the user had modified anything in the customization window, then maybe it could show a warning message if the user alters the fields in the main GUI.

Glitch #3
The customization GUI has the "Cancel" button as active.  It should be the "Save" button which is active.  Currently hitting the enter key causes the GUI to cancel, not save.