Fixed the following bug reports from Danny:
BUG #1
I just tested it with log checked and Start_Value -1.  This causes seadas to freeze up, at least in IntelliJ.

GLITCH/ANNOYANCE #1
Also I had an idea for the latest change where we grey out the contour group.  It would be better to grey it out only if the user has clicked the "Save" button.  If the user clicks "Preview/Edit" and then just views it and clicks "Cancel" there is no reason to grey it out as the user has not made changes.  Greying out on "Cancel" can get a little annoying if the user is trying out or learning the different auto-create possiblities as they keep having to add and delete groups.